implemented repository in player list view model

+++ app/src/main/java/com/example/the7wonders/data/repository/PlayerRepositoryImpl.kt
@@ -22,6 +22,8 @@ class PlayerRepositoryImpl @Inject constructor(private val playerDao: PlayerDao)
     }
 
     override suspend fun deletePlayer(player: PlayerModel) {
+        //TODO("Add actual error handling")
+        if (player.id == null) return
         playerDao.deletePlayer(player.id)
     }
